+++ src/KeyValueStoreDiscovery.php
@@ -230,13 +230,6 @@ class KeyValueStoreDiscovery extends AbstractEditableDiscovery
             $this->loadBindingsForKey($key);
         }
 
-        // Ignore duplicates
-        foreach ($this->bindingsByKey[$key] as $other) {
-            if ($binding->equals($other)) {
-                return;
-            }
-        }
-
         $this->initializeBinding($binding);
 
         $this->bindingsByKey[$key][] = $binding;